sem2/Lab01/requirements.txt
PyQt5
msgpack
pycryptodome